requirements.txt
fastapi
uvicorn
pymongo
pydantic
python-jose[cryptography]
passlib[bcrypt]